updated pca code to work to data_dict and cleared output for all other code

+++ utils.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 def load_h5(fpath):
     data_h5file = h5py.File(fpath, 'r')
     # load a snippit of data and get rid of un-needed singleton dimensions
-    data_snip = np.squeeze(np.array(data_h5file['imaging'], dtype=int))
+    data_snip = np.squeeze(np.array(data_h5file['imaging'], dtype=np.uint16))
 
     """ typically it's good to have time as the last dimension because one doesn't usually iterate through time, so we'll
      reorganize the data dimension order"""
